FEAT -- Rename files names and add more files to .gitignore
